+++ lib/templates.ts
@@ -13,62 +13,6 @@ import { validateTemplate } from "@lib/utils/form-builder/validate";
 import { dateHasPast } from "@lib/utils";
 import { validateTemplateSize } from "@lib/utils/validateTemplateSize";
 
-export async function updateFormSaveAndResume(
-  formID: string,
-  saveAndResume: boolean
-): Promise<FormRecord | null> {
-  const { user } = await authorization.canEditForm(formID).catch((e) => {
-    logEvent(
-      e.user.id,
-      { type: "Form", id: formID },
-      "AccessDenied",
-      AuditLogAccessDeniedDetails.AccessDenied_AttemptToSetSaveAndResume
-    );
-    throw e;
-  });
-
-  const updatedTemplate = await prisma.template
-    .update({
-      where: {
-        id: formID,
-      },
-      data: {
-        saveAndResume: saveAndResume ?? false,
-      },
-      select: {
-        id: true,
-        created_at: true,
-        updated_at: true,
-        name: true,
-        jsonConfig: true,
-        isPublished: true,
-        deliveryOption: true,
-        securityAttribute: true,
-        formPurpose: true,
-        publishDesc: true,
-        publishFormType: true,
-        publishReason: true,
-        saveAndResume: true,
-        notificationsInterval: true,
-      },
-    })
-    .catch((e) => {
-      return prismaErrors(e, null);
-    });
-
-  if (updatedTemplate === null) return updatedTemplate;
-
-  logEvent(
-    user.id,
-    { type: "Form", id: formID },
-    AuditLogEvent.ChangeFormSaveAndResume,
-    AuditLogDetails.SetSaveAndResume,
-    { saveAndResume: saveAndResume ? "On" : "Off" }
-  );
-
-  return _parseTemplate(updatedTemplate);
-}
-
 /**
  * Remove DeliveryOption from template. Form responses will be sent to the Vault.
  * @param formID The unique identifier of the form you want to modify
